+++ compiam/rhythm/__init__.py
@@ -2,7 +2,7 @@ import sys
 
 from compiam.utils import get_tool_list
 
-from compiam.rhythm.tabla_transcription import FourWayTabla
+from compiam.rhythm.transcription import FourWayTabla
 from compiam.rhythm.akshara_pulse_tracker import AksharaPulseTracker
 
 # Show user the available tools
