+++ compiam/rhythm/transcription/model.py
